pyproject.toml
[tool.poetry.dependencies]
beautifulsoup4 = "4.11.1"
build = "0.8.0"
cachetools = "5.2.0"
certifi = "2022.6.15"
charset-normalizer = "2.0.12"
idna = "3.3"
loguru = "0.6.0"
oauthlib = "3.2.0"
pathlib = "1.0.1"
packaging = "21.3"
pep517 = "0.12.0"
protobuf = "4.21.1"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pyparsing = "3.0.9"
requests = "2.28.0"
requests-oauthlib = "1.3.1"
rsa = "4.8"
six = "1.16.0"
soupsieve = "2.3.2.post1"
toml = "0.10.2"
tomli = "2.0.1"
uritemplate = "4.1.1"
urllib3 = "1.26.9"
python-dotenv = "0.20.0"